Ensure sudo in scripts

+++ link.sh
@@ -2,6 +2,12 @@
 
 set -eu
 
+if [[ $EUID -eq 0 ]]; then
+  echo "This script should not be run as root" >&2
+  exit 1
+fi
+sudo -v
+
 mkdir -p $HOME/.config
 
 rm -rf $HOME/.config/nvim
